log in to edit product to invalidate cache
